refactor(ui): migre pure-onf-dashboard en tsx
